Indent "elif" line with "then" on the same line properly
If "elif" line has "then" on the same line, the line is indented less.
This is caused by doubly indented out by "outc" and "else_case" checkings.

Now the search pattern for "else_case" is modified so "elif" without "then"
is indented properly. Also added "\b" to check the keywords exactly.

And test4 is added to check "if"/"else"/"elif" blocks are indented properly.