Handle quoted here document delimeter
A quoted here-doc delimeter (like <<"END" or <<'END') is used to avoid
shell expansion in the here document.

But because quoted strings are removed in "test_record", the existing
pattern match does not extract the delimeter name, which causes bad output.

Now the quotes from here-doc quoted delimeter are removed before collapsing
quotes.